Redirect in list, Fixed dropdown, New Purple
Changes:

Fix: Fixed the drop-down menu. On some resolutions before you could see a pixel or more of the redirect checkbox
Chore: Updated the images in README
Feature: Started working on Mobile UI (Disabled because its not finished)
Feature: Better formatting of the date in the list
Fix: Updated the colors of the new purple to the favicon.
Fix: Removed the animation for the redirect button when editing an endpoint.